requirements.txt
fastapi
uvicorn
pdfminer.six
sentence-transformers
faiss-cpu
numpy
requests
python-multipart